BACKEND/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
fastapi==0.110.1
Flask==3.0.3
flask-restx==1.3.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
restx==0.1.14
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
uuid==1.30
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
